transfer api function to diff js file
